norm:split_withquote
